+++ src/test/groovy/ru/vyarus/gradle/plugin/lib/ReportsAggregationKitTest.groovy
@@ -19,7 +19,7 @@ class ReportsAggregationKitTest extends AbstractKitTest {
             }
                 
             javaLib {
-                aggregateSubprojectReports()
+                aggregateReports()
             }
             
             allprojects {