+++ CHANGELOG.md
@@ -14,7 +14,7 @@
     - enableSnapshotsSigning() - enables signing for snapshot versions (disabled by default)
     - pom - shortcut for pom plugin configuration (to use instead of pomGeneration)
     - autoModuleName - shortcut for defining Automatic-Module-Name manifest property
-    - aggregateSubprojectReports() - supposed to be used in the root project to aggregate
+    - aggregateReports() - supposed to be used in the root project to aggregate
        test reports and jacoco coverage (applied with "base" plugin, adds test and jacocoTestReport tasks)